API calls for delete/update/get TerrainObject/Program added and runProgram API call updated

+++ src/main/java/com/example/hamster_backend/service/impl/ProgramServiceImpl.java
@@ -1,13 +1,14 @@
 package com.example.hamster_backend.service.impl;
 
 import com.example.hamster_backend.model.entities.Program;
-import com.example.hamster_backend.model.entities.User;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
+import com.example.hamster_backend.model.session.AuthorizationUtils;
 import com.example.hamster_backend.repositories.ProgramRepository;
 import com.example.hamster_backend.service.ProgramService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.Optional;
 import java.util.Set;
 
@@ -20,7 +21,7 @@ public class ProgramServiceImpl implements ProgramService {
     @Override
     public void compareAndUpdateDatabase(Program program) {
         program.setHashValue(program.hashCode());
-        Set<Program> userPrograms = programRepository.findAllByUserID(program.getUserID());
+        Set<Program> userPrograms = programRepository.findAllByUserId(program.getUserId());
 
         if (!userPrograms.isEmpty()) {
             userPrograms.forEach(t -> {
@@ -34,19 +35,77 @@ public class ProgramServiceImpl implements ProgramService {
         }
     }
 
-    @Override
-    public void saveOrUpdate(Program program) {
-        User u = ((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal());
-        program.setUserID(u.getId());
-        compareAndUpdateDatabase(program);
-    }
 
     @Override
     public void delete(long programId) {
         Optional<Program> o = programRepository.findById(programId);
-        if(o.isPresent()){
+        if (o.isPresent()) {
             Program p = o.get();
-            programRepository.delete(p);
+            if (AuthorizationUtils.getAuthUserId().equals(p.getUserId())) {
+                programRepository.delete(p);
+            }
         }
+    }
+
+    @Override
+    public Set<Program> getProgramBasicData(long id) {
+        return programRepository.findAllProgramBasicDataByUserID(id);
+    }
+
+    @Override
+    public Program getProgram(long id) {
+        return programRepository.findProgramByProgramId(id);
+    }
+
+    @Override
+    public ArrayList<Program> getProgramsByNames(ArrayList<String> names) {
+        ArrayList<Program> programs = new ArrayList<>();
+        names.forEach(n -> {
+            programs.add(programRepository.findProgramByProgramName(n));
+        });
+        return programs;
+    }
+
+    @Override
+    public Set<Program> getAllNeededProgramToRun(String mainFileCode) {
+        Set<String> foundClasses = new HashSet<>();
+        Set<Program> programs = new HashSet<>();
+        Set<String> classNames = new HashSet<>(Program.extractImportedClasses(mainFileCode));
+
+        while (!classNames.isEmpty()) {
+            classNames.forEach(n -> {
+                Program program = programRepository.findProgramByProgramName(n);
+                classNames.addAll(Program.extractImportedClasses(program.getSourceCode()));
+                foundClasses.add(n);
+            });
+            classNames.removeIf(foundClasses::contains);
+        }
+
+        foundClasses.forEach(n -> programs.add(programRepository.findProgramByProgramName(n)));
+
+        return programs;
+    }
+
+    @Override
+    public void save(Program program) {
+        Optional<Program> o = programRepository.findById(program.getProgramId());
+        if (o.isPresent()) {
+            Program p = o.get();
+            if (AuthorizationUtils.getAuthUserId().equals(p.getUserId())) {
+                programRepository.update(p.getProgramId(), p.getSourceCode());
+            }
+        } else {
+            programRepository.save(program);
+        }
+    }
+
+    @Override
+    public void updatePath(Program program) {
+        programRepository.updatePath(program.getProgramId(), program.getProgramPath());
+    }
+
+    @Override
+    public void updateName(Program program) {
+        programRepository.updateName(program.getProgramId(), program.getProgramName());
     }
 }
